Fixed a documentation error regarding Multiple insert (#737)
the current docs show that you can pass each row to be inserted as additional params to the query method, however when you try to do that it errors.  In looking up this error here: https://forum.nette.org/en/21412-nette-database-insert-multiple-rows-in-one-query I realized that the multiple rows need to be passed as an array in the second argument which solved the problem.