Update home-page.texy: fix DB connection code
When using a code for `app/Presenters/HomepagePresenter.php` from the documentation, it gave an error:
`syntax error, unexpected 'private' (T_PRIVATE), expecting variable (T_VARIABLE)`
Because indeed the code was corrupt for some reason.
Also checked - documentation in other languages also contains this error in the code for `app/Presenters/HomepagePresenter.php`. So all others need to be fixed as well.