Add info - creating ajax request in the browser
Hi,

I send a proposal to complete info about the http ajax request. The introduction paragraph talks about an ajax http request on the server, but newcomers need a brief input on how an ajax request is actually created - that's what the paragraph should ideally start with. It is also appropriate to give an Ajax request as such a separate title.